+++ clients/client.py
@@ -63,10 +63,10 @@ def writeLastTime():
 
 
 class Chooser(QDialog):
-    def __init__(self, parent):
+    def __init__(self, Qapp):
         self.APIVersion = Plom_API_Version
         super(Chooser, self).__init__()
-        self.parent = parent
+        self.parent = Qapp
         print(
             "Plom Client {} (communicates with api {})".format(
                 __version__, self.APIVersion
@@ -125,10 +125,7 @@ class Chooser(QDialog):
         try:
             messenger.startMessenger()
         except PlomBenignException as e:
-            ErrorMessage(
-                "Could not connect to server.\n\n"
-                "{}".format(e)
-            ).exec_()
+            ErrorMessage("Could not connect to server.\n\n" "{}".format(e)).exec_()
             return
 
         try:
@@ -143,6 +140,22 @@ class Chooser(QDialog):
         except PlomAuthenticationException as e:
             ErrorMessage("Could not authenticate: {}".format(e)).exec_()
             return
+        except PlomExistingLoginException as e:
+            if (
+                SimpleMessage(
+                    "You appear to be already logged in!\n\n"
+                    "  * Perhaps a previous session crashed?\n"
+                    "  * Do you have another client running,\n"
+                    "    e.g., on another computer?\n\n"
+                    "Should I force-logout the existing authorisation?"
+                    " (and then you can try to log in again)\n\n"
+                    "The other client will likely crash."
+                ).exec_()
+                == QMessageBox.Yes
+            ):
+                messenger.clearAuthorisation(user, pwd)
+            return
+
         except PlomSeriousException as e:
             ErrorMessage(
                 "Could not get authentication token.\n\n"
@@ -157,7 +170,7 @@ class Chooser(QDialog):
             v = self.getv()
             self.setEnabled(False)
             self.hide()
-            markerwin = marker.MarkerClient()
+            markerwin = marker.MarkerClient(self.parent)
             markerwin.my_shutdown_signal.connect(self.on_marker_window_close)
             markerwin.show()
             markerwin.getToWork(messenger, pg, v, lastTime)
@@ -258,40 +271,41 @@ class Chooser(QDialog):
             return
         mport = self.ui.mportSB.value()
         # save those settings
-        #self.saveDetails()   # TODO?
+        # self.saveDetails()   # TODO?
 
         # TODO: might be nice, but needs another thread?
-        #self.ui.infoLabel.setText("connecting...")
-        #self.ui.infoLabel.repaint()
+        # self.ui.infoLabel.setText("connecting...")
+        # self.ui.infoLabel.repaint()
 
         # Have Messenger login into to server
         messenger.setServerDetails(server, mport)
         try:
             r = messenger.startMessenger()
         except PlomBenignException as e:
-            ErrorMessage(
-                "Could not connect to server.\n\n"
-                "{}".format(e)
-            ).exec_()
+            ErrorMessage("Could not connect to server.\n\n" "{}".format(e)).exec_()
             return
         self.ui.infoLabel.setText(r)
 
         info = messenger.getInfoGeneral()
-        self.ui.markGBox.setTitle('Marking information for “{}”'.format(info["testName"]))
+        self.ui.markGBox.setTitle(
+            "Marking information for “{}”".format(info["testName"])
+        )
         pg = self.getpg()
         v = self.getv()
         self.ui.pgSB.setVisible(False)
         self.ui.vSB.setVisible(False)
 
         self.ui.vDrop.clear()
-        self.ui.vDrop.addItems([str(x+1) for x in range(0, info["numVersions"])])
+        self.ui.vDrop.addItems([str(x + 1) for x in range(0, info["numVersions"])])
         if v:
             if v >= 1 and v <= info["numVersions"]:
                 self.ui.vDrop.setCurrentIndex(v - 1)
         self.ui.vDrop.setVisible(True)
 
         self.ui.pgDrop.clear()
-        self.ui.pgDrop.addItems(["Q{}".format(x+1) for x in range(0, info["numGroups"])])
+        self.ui.pgDrop.addItems(
+            ["Q{}".format(x + 1) for x in range(0, info["numGroups"])]
+        )
         if pg:
             if pg >= 1 and pg <= info["numGroups"]:
                 self.ui.pgDrop.setCurrentIndex(pg - 1)
@@ -299,6 +313,11 @@ class Chooser(QDialog):
         # TODO should we also let people type in?
         self.ui.pgDrop.setEditable(False)
         self.ui.vDrop.setEditable(False)
+        # put focus at username or password line-edit
+        if len(self.ui.userLE.text()) > 0:
+            self.ui.passwordLE.setFocus(True)
+        else:
+            self.ui.userLE.setFocus(True)
 
     @pyqtSlot(int)
     def on_other_window_close(self, value):