+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * More robust networking by moving to https (rather than raw sockets). This is handled by the "requests" library on the client side and the "aiohttp" library on the server side.
 * Client: config and saved comments now more human-readable in toml files.
 * Client: can download test/server info before logging in.
+* Client is more pessimistic about errors and will crash instead of trying to continue
+in most cases.	
+* Client checks for double-logins and can force logout a previous session.
 
 ### Fixed
 * many fixes, especially related to client crashes due to networking.